Fix script create icons command, moves sidebar to component folder
